Players now move around the Board! Currently: First one back to the start wins!
Full Changelog:
* Players now store the 'last position' they were standing at
* Board objects now return a NullTile if they're asked for a tile out-of-bounds
* Hand.getCastableSpells() returns the spells someone can cast (currently nothing)
* Tiles now have temporary, unimplemented onPass() and onLand() functions
* Match Objects now ask the appropriate PlayerRepresentative what spell that Player would like to cast at the start of their turn
* Added BoardIterator!
The BoardIterator moves Players on their turn, asking the respective PlayerRepresentative for the number they've rolled on the dice. Then, for each move on the dice they have yet to make, the BoardIterator figures out which directions they can move in and asks the player which way they would like to go. If the Player cannot go in a particular direction, the BoardIterator does not provide that option. If the Player can only go in one direction, the BoardIterator moves them automatically. The BoardIterator calls Tile.onPass() and Tile.onLand() as appropriate, and calls Match.updateObservers() after each move.