* PlayerRepresentatives now return the Position of a selected Tile rather than the Tile itself.
This was done due to limitations of the networking features in the Online_Multiplayer branch. When PlayerReps return Tiles, online players receive Tiles that the game doesn't not recognize as already being on the board. When the game tries to use or alter these Tiles in any way, these changes aren't reflected onto the intended Tile on the Board, therefore failing to alter the intended Tile (but only for players over the internet).
By passing integer positions, we overcome this limitation at the expense of additional Board lookups, and Commands needing to hold additional references to the Board to perform these lookups.